requirements.txt
APScheduler==3.7.0
asgiref==3.4.1
certifi==2021.5.30
chardet==4.0.0
Django==3.2.5
djangorestframework==3.12.4
idna==2.10
lxml==4.6.3
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
tzlocal==2.1
urllib3==1.26.6
